Extract part of credentials to Identity module
