+++ sdk/core/azure-core/src/credentials.cpp
@@ -1,11 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include "azure/core/credentials/policy/policies.hpp"
+#include <azure/core/credentials.hpp>
 
-using namespace Azure::Core::Credentials::Policy;
+using namespace Azure::Core;
 
-std::unique_ptr<Azure::Core::Http::RawResponse> BearerTokenAuthenticationPolicy::Send(
+std::unique_ptr<Http::RawResponse> Azure::Core::BearerTokenAuthenticationPolicy::Send(
     Context const& context,
     Http::Request& request,
     Http::NextHttpPolicy policy) const
