+++ sdk/identity/CHANGELOG.md
@@ -1,0 +1,6 @@
+# Release History
+
+## 1.0.0-beta.3 (Unreleased)
+
+* Support for Client Secret Credential
+* Support for Environment Credential
