+++ sdk/identity/azure-identity/src/client_secret_credential.cpp
@@ -1,18 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include "azure/core/credentials/credentials.hpp"
-#include "azure/core/http/body_stream.hpp"
-#include "azure/core/http/curl/curl.hpp"
-#include "azure/core/http/http.hpp"
-#include "azure/core/http/pipeline.hpp"
+#include <azure/identity/client_secret_credential.hpp>
 
-#include <cstdlib>
 #include <iomanip>
 #include <sstream>
-#include <stdexcept>
 
-using namespace Azure::Core::Credentials;
+using namespace Azure::Identity;
 
 namespace {
 std::string UrlEncode(std::string const& s)
@@ -39,11 +33,11 @@ std::string UrlEncode(std::string const& s)
 }
 } // namespace
 
-std::string const Azure::Core::Credentials::ClientSecretCredential::g_aadGlobalAuthority
+std::string const Azure::Identity::ClientSecretCredential::g_aadGlobalAuthority
     = "https://login.microsoftonline.com/";
 
-AccessToken Azure::Core::Credentials::ClientSecretCredential::GetToken(
-    Context const& context,
+Azure::Core::AccessToken Azure::Identity::ClientSecretCredential::GetToken(
+    Azure::Core::Context const& context,
     std::vector<std::string> const& scopes) const
 {
   static std::string const errorMsgPrefix("ClientSecretCredential::GetToken: ");
@@ -187,81 +181,16 @@ AccessToken Azure::Core::Credentials::ClientSecretCredential::GetToken(
             + std::chrono::seconds(expiresInSeconds < 0 ? 0 : expiresInSeconds),
     };
   }
-  catch (AuthenticationException const&)
+  catch (Azure::Core::AuthenticationException const&)
   {
     throw;
   }
   catch (std::exception const& e)
   {
-    throw AuthenticationException(e.what());
+    throw Azure::Core::AuthenticationException(e.what());
   }
   catch (...)
   {
-    throw AuthenticationException("unknown error");
+    throw Azure::Core::AuthenticationException("unknown error");
   }
-}
-
-Azure::Core::Credentials::EnvironmentCredential::EnvironmentCredential()
-{
-#ifdef _MSC_VER
-#pragma warning(push)
-// warning C4996: 'getenv': This function or variable may be unsafe. Consider using _dupenv_s
-// instead.
-#pragma warning(disable : 4996)
-#endif
-
-  auto tenantId = std::getenv("AZURE_TENANT_ID");
-  auto clientId = std::getenv("AZURE_CLIENT_ID");
-
-  auto clientSecret = std::getenv("AZURE_CLIENT_SECRET");
-  auto authority = std::getenv("AZURE_AUTHORITY_HOST");
-
-  // auto username = std::getenv("AZURE_USERNAME");
-  // auto password = std::getenv("AZURE_PASSWORD");
-  //
-  // auto clientCertificatePath = std::getenv("AZURE_CLIENT_CERTIFICATE_PATH");
-
-#ifdef _MSC_VER
-#pragma warning(pop)
-#endif
-
-  if (tenantId != nullptr && clientId != nullptr)
-  {
-    if (clientSecret != nullptr)
-    {
-      if (authority != nullptr)
-      {
-        m_credentialImpl.reset(
-            new ClientSecretCredential(tenantId, clientId, clientSecret, authority));
-      }
-      else
-      {
-        m_credentialImpl.reset(new ClientSecretCredential(tenantId, clientId, clientSecret));
-      }
-    }
-    // TODO: These credential types are not implemented. Uncomment when implemented.
-    // else if (username != nullptr && password != nullptr)
-    //{
-    //  m_credentialImpl.reset(
-    //      new UsernamePasswordCredential(username, password, tenantId, clientId));
-    //}
-    // else if (clientCertificatePath != nullptr)
-    //{
-    //  m_credentialImpl.reset(
-    //      new ClientCertificateCredential(tenantId, clientId, clientCertificatePath));
-    //}
-  }
-}
-
-AccessToken Azure::Core::Credentials::EnvironmentCredential::GetToken(
-    Context const& context,
-    std::vector<std::string> const& scopes) const
-{
-  if (!m_credentialImpl)
-  {
-    throw AuthenticationException("EnvironmentCredential authentication unavailable. "
-                                  "Environment variables are not fully configured.");
-  }
-
-  return m_credentialImpl->GetToken(context, scopes);
 }
