+++ sdk/identity/azure-identity/version.txt
@@ -1,0 +1,1 @@
+1.0.0-beta.2
