+++ sdk/identity/ci.yml
@@ -1,0 +1,37 @@
+# NOTE: Please refer to https://aka.ms/azsdk/engsys/ci-yaml before editing this file.
+trigger:
+  branches:
+    include:
+      - master
+      - feature/*
+      - release/*
+      - hotfix/*
+  paths:
+    include:
+      - eng/
+      - CMakeLists.txt
+      - sdk/core
+      - sdk/identity
+
+pr:
+  branches:
+    include:
+      - master
+      - feature/*
+      - release/*
+      - hotfix/*
+  paths:
+    include:
+      - eng/
+      - CMakeLists.txt
+      - sdk/core
+      - sdk/identity
+
+stages:
+  - template: ../../eng/pipelines/templates/stages/archetype-sdk-client.yml
+    parameters:
+      ServiceDirectory: identity
+      CtestRegex: azure-identity
+      Artifacts:
+        - Name: azure-identity
+          Path: azure-identity
