+++ sdk/storage/azure-storage-blobs/inc/azure/storage/blobs/blob_batch_client.hpp
@@ -121,7 +121,7 @@ namespace Azure { namespace Storage { namespace Blobs {
      */
     explicit BlobBatchClient(
         const std::string& serviceUri,
-        std::shared_ptr<Core::Credentials::TokenCredential> credential,
+        std::shared_ptr<Identity::TokenCredential> credential,
         const BlobBatchClientOptions& options = BlobBatchClientOptions());
 
     /**
