+++ sdk/storage/ci.yml
@@ -11,6 +11,7 @@ trigger:
       - eng/
       - CMakeLists.txt
       - sdk/core
+      - sdk/identity
       - sdk/storage
 
 pr:
@@ -25,6 +26,7 @@ pr:
       - eng/
       - CMakeLists.txt
       - sdk/core/
+      - sdk/identity/
       - sdk/storage
 
 stages:
